+++ src/Icons/tests/Unit/Twig/UXIconRuntimeTest.php
@@ -16,6 +16,7 @@ use Psr\Log\LoggerInterface;
 use Symfony\UX\Icons\Exception\IconNotFoundException;
 use Symfony\UX\Icons\IconRendererInterface;
 use Symfony\UX\Icons\Twig\UXIconRuntime;
+use Twig\Extra\Html\HtmlAttr\AttributeValueInterface;
 
 /**
  * @author Simon André <smn.andre@gmail.com>
@@ -39,5 +40,35 @@ class UXIconRuntimeTest extends TestCase
         $runtime = new UXIconRuntime($renderer, false);
         $this->expectException(IconNotFoundException::class);
         $runtime->renderIcon('not_found');
+    }
+
+    public function testRenderIconConvertsTypedAttributeValues()
+    {
+        $renderer = $this->createMock(IconRendererInterface::class);
+        $renderer->expects($this->once())
+            ->method('renderIcon')
+            ->with('user', ['class' => 'size-4', 'data-action' => 'click->menu#toggle'])
+            ->willReturn('<svg></svg>');
+
+        $runtime = new UXIconRuntime($renderer);
+        $this->assertSame('<svg></svg>', $runtime->renderIcon('user', [
+            'class' => 'size-4',
+            'data-action' => $this->createTypedValue('click->menu#toggle'),
+            'aria-hidden' => $this->createTypedValue(null),
+        ]));
+    }
+
+    private function createTypedValue(?string $value): AttributeValueInterface
+    {
+        return new class($value) implements AttributeValueInterface {
+            public function __construct(private readonly ?string $value)
+            {
+            }
+
+            public function getValue(): ?string
+            {
+                return $this->value;
+            }
+        };
     }
 }